Update LanguageRedirectMiddleware.php
Reading typoscript settings like this

[traverse(page, "uid") in [968,967,202,965,969]]
    config.tx_locate = 0
[ELSE]
    config.tx_locate = 1
[GLOBAL]

does not work reliable with BackendConfigurationManager. If we use the ConfigurationManager to read frontend typoscript config, it works as expected.

As far as I see, the BackendConfigurationManager is not really supposed to read frontend typoscript code anyway. See comment at the start of the class:

/*
 * Load TypoScript of a page in backend mode.
 *
 * Extbase Backend modules can be configured with Frontend TypoScript. This is of course a very
 * bad thing, but it is how it is ^^ (we'll get rid of this at some point, promised!)
*/